+++ .eslintrc.js
@@ -4,6 +4,8 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
+const path = require('path')
+
 module.exports = {
   env: {
     browser: true,
@@ -29,10 +31,17 @@ module.exports = {
         aspects: ['invalidHref', 'preferButton']
       }
     ],
+    'max-len': [
+      'error',
+      {
+        code: 150, // Effectively disable this rule and allow Prettier to handle it
+        comments: 100
+      }
+    ],
     'notice/notice': [
       'error',
       {
-        templateFile: '.copyright.js'
+        templateFile: path.join(__dirname, '.copyright.js')
       }
     ],
     'space-before-function-paren': [