+++ .stylelintrc.js
@@ -2,5 +2,8 @@ module.exports = {
   extends: ['stylelint-config-carbon'],
   allowEmptyInput: true,
   reportNeedlessDisables: true,
-  reportInvalidScopeDisables: true
+  reportInvalidScopeDisables: true,
+  rules: {
+    'selector-pseudo-class-no-unknown': null
+  }
 }