+++ scripts/micromanage-scripts/utils.js
@@ -7,6 +7,12 @@
 const { execSync } = require('child_process')
 const path = require('path')
 
+/**
+ * Execute a command line command.
+ *
+ * @param {string} cmd Command to execute.
+ * @returns {string} Output of the command.
+ */
 function exec(cmd) {
   return execSync(cmd, {
     env: {
@@ -17,15 +23,44 @@ function exec(cmd) {
     .trim()
 }
 
-function getFiles(dir, extensions) {
+/**
+ * Given a file, find the package under which the file resides.
+ *
+ * @param {string} f File path.
+ * @returns {object} Package object of the package containing the file.
+ */
+function getPackageForFile(f) {
+  const packages = getPackages()
+
+  return packages.find((pkg) => f.startsWith(pkg.path))
+}
+
+/**
+ * Get all of the files in a particular directory, given some filters using the `find` utility.
+ *
+ * @param {string} dir Directory in which to look.
+ * @param {Array} extensions File extensions to include.
+ * @param {Array} [exclusions] Specific paths to exclude, formatted the way `find` expects for
+ * `-not -path <path>` entries.
+ * @returns {Array} Array of files.
+ */
+function getFiles(dir, extensions, exclusions = []) {
+  exclusions = exclusions.map((exclusion) => `-not -path "${exclusion}"`)
+  const exclusionsString = exclusions.join(' ')
+
   return extensions
     .map((extension) => {
-      const files = exec(`find ${dir} -name "*${extension}" -type f`)
+      const files = exec(`find ${dir} -name "*${extension}" -type f ${exclusionsString}`)
       return files.length === 0 ? [] : files.split('\n')
     })
     .reduce((prev, cur) => [...prev, ...cur], [])
 }
 
+/**
+ * Get an object for each package/workspace, as defined in the top-level package.json file.
+ *
+ * @returns {Array} Array of package info objects.
+ */
 function getPackages() {
   const packageJson = require(path.join(process.cwd(), 'package.json'))
 
@@ -37,6 +72,7 @@ function getPackages() {
     return {
       name: p.name,
       dependencies: p.dependencies,
+      devDependencies: p.devDependencies,
       path: packagePath,
       private: !!p.private,
       version: p.version
@@ -44,6 +80,11 @@ function getPackages() {
   })
 }
 
+/**
+ * Get a list of all git tags, sorted by taggerdate.
+ *
+ * @returns {Array} Array of tags.
+ */
 function getTags() {
   const tagsCommandOutput = exec('git tag --sort=taggerdate')
   return tagsCommandOutput.split('\n')
@@ -52,6 +93,7 @@ function getTags() {
 module.exports = {
   exec,
   getFiles,
+  getPackageForFile,
   getPackages,
   getTags
 }