+++ services/web-app/icons/24Cloud.js
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 /**
  * Copyright IBM Corp. 2021, 2021
  *