+++ services/web-app/next.config.js
@@ -13,6 +13,9 @@ module.exports = {
     locales: ['en-US'],
     defaultLocale: 'en-US'
   },
+  images: {
+    domains: ['raw.githubusercontent.com']
+  },
   sassOptions: {
     includePaths: [path.join(__dirname, 'styles')],
     prependData: `
@@ -24,6 +27,7 @@ module.exports = {
       @use '~@carbon/react/scss/zone';
     `
   },
+  swcMinify: true,
   webpack(config) {
     const rules = config.module.rules
       .find((rule) => typeof rule.oneOf === 'object')
@@ -32,6 +36,7 @@ module.exports = {
     rules.forEach((rule) => {
       rule.use.forEach((moduleLoader) => {
         if (
+          moduleLoader.loader &&
           moduleLoader.loader.includes('css-loader') &&
           typeof moduleLoader.options.modules === 'object'
         ) {