+++ services/web-app/pages/index.js
@@ -9,7 +9,7 @@ import { useContext, useEffect } from 'react'
 
 import { LayoutContext } from '@/layouts/layout'
 import { NextSeo } from 'next-seo'
-import { Svg16Carbon } from '@/icons/index.js'
+import { Svg16Carbon } from '@carbon-platform/icons'
 import defaultSeo from '@/config/seo.json'
 
 const navData = [